Moved the agreement interfaces
