+++ client/src/interfaces/agreements/agreementView.ts
@@ -1,43 +1,3 @@
-export interface AgreementInList {
-	AgreementId: number;
-	CreditCardNo: string;
-	ReserveId: number | null;
-	VehicleId: number;
-	MakeId: number;
-	VehicleNo: string;
-	VehicleMakeName: string;
-	ModelId: number;
-	ModelName: string;
-	Year: number;
-	Trim: string | null;
-	LicenseNo: string;
-	CheckoutDate: string;
-	CheckinDate: string;
-	ReturnDate: string;
-	CreatedDate: string;
-	CreatedByName: string;
-	CustomerId: number;
-	FirstName: string;
-	LastName: string;
-	PONo: string | null;
-	RONo: string | null;
-	HPhone: string | null;
-	CPhone: string | null;
-	BPhone: string | null;
-	AgreementStatusId: number;
-	AgreementNumber: string;
-	TotalAmount: number;
-	Note: string | null;
-	Color: string;
-	AgreementStatusName: string;
-	CheckoutLocationId: number;
-	CheckinLocationId: number;
-	CheckoutLocationName: string;
-	CheckinLocationName: string;
-	VehicleTypeId: number;
-	VehicleType: string;
-}
-
 export interface AgreementDataFull {
 	agreementId: number;
 	agreementNumber: string;
@@ -228,50 +188,4 @@ export interface AgreementDataFull {
 		forfeitureAmount: number;
 		totalCanceledAmount: number;
 	};
-}
-
-export interface AgreementSummary {
-	baseRate: number;
-	totalDays: number;
-	miscCharges: [];
-	totalMiscChargesTaxable: number;
-	totalMiscChargesNonTaxable: number;
-	subTotal: number;
-	totalTax: number;
-	dropOffCharges: number;
-	total: number;
-	promotionDiscount: number;
-	rateSummaryItems: {
-		type: string;
-		rate: number;
-		units: number;
-		startDate: string;
-		endDate: string;
-		kMorMileageAllowed: number;
-	}[];
-	avgPerDayRate: number;
-	taxes: {
-		taxId: number;
-		locationTaxId: number;
-		name: string;
-		description: string;
-		value: number;
-		isOption: false;
-		isDeleted: false;
-		isDefaultSelectedForReservation: boolean;
-		misChargeId: string | null;
-	}[];
-	promotions: null;
-	maxTotalDays: number;
-	freeMiles: number;
-	extraMilesCharge: number;
-	initialCharge: number;
-	initialChargeHours: number;
-	avgPerHourRate: number;
-	advancePayment: number | null;
-	balanceDue: number;
-	inventoryCharges: number | null;
-	totalInventoryChargesTaxable: number;
-	totalInventoryChargesNonTaxable: number;
-	finalBaseRate: number;
 }
