+++ client/src/redux/thunks/searchAgreementsThunks.ts
@@ -1,7 +1,7 @@
 import { createAsyncThunk } from "@reduxjs/toolkit";
 import axios from "axios";
 import appAxiosInstance from "../../api/appAxiosInstance";
-import { AgreementInList } from "../../interfaces/agreement";
+import { AgreementInList } from "../../interfaces/agreements/agreementSearch";
 import { RootState } from "../store";
 
 export const fetchAgreementsThunk = createAsyncThunk<
