+++ client/src/utils/fakeData.ts
@@ -1,4 +1,6 @@
-import { AgreementInList, AgreementSummary, AgreementDataFull } from "../interfaces/agreement";
+import { AgreementInList } from "../interfaces/agreements/agreementSearch";
+import { AgreementDataFull } from "../interfaces/agreements/agreementView";
+import { AgreementSummary } from "../interfaces/agreements/agreementSummary";
 import { CustomersInList } from "../interfaces/customers";
 import { ReservationsInList } from "../interfaces/reservations/reservationSearch";
